make codemirror component
