+++ src/app/main/apps/home/widgets/Widget1.js
@@ -2,8 +2,6 @@ import { makeStyles} from '@material-ui/core/styles';
 import Typography from '@material-ui/core/Typography';
 import React from 'react';
 import ButtonBase from '@material-ui/core/ButtonBase';
-import { MemoryRouter as Router } from 'react-router';
-import { Link as RouterLink } from 'react-router-dom';
 import Link from '@material-ui/core/Link';
